apply cfp's patch
